Got rid of some useless code and inlined TrackNode.fields property method as there was only one use. Created a superclass for each AlbumNode and TrackNode called Node that supplies total ordering so that things are sorted nicely now. Changed selection of parent to match selection of children so that it's similarly fast (select a whole row at once rather than select each row/column cell.
One issue exists now where I have two albums with the same name. The bisect
search finds the first occurance and (de)selects the wrong album. I'll either
need to provide a key to bisect in the AlbumNode.rowOfChild method, or maybe put
albums with the same name together. I'll sleep on it and come back with a
solution.